mask: use custom outerJoin to try and preserve partitioner https://github.com/Open-EO/openeo-geotrellis-extensions/issues/424
